Initial single-threaded echo server
